+++ src/Config/Resilient.php
@@ -5,13 +5,21 @@ namespace Innmind\OperatingSystem\Config;
 
 use Innmind\OperatingSystem\Config;
 use Innmind\HttpTransport\Transport;
+use Innmind\Time\Period;
+use Innmind\Immutable\Sequence;
 
 /**
  * @psalm-immutable
  */
-enum Resilient
+final class Resilient
 {
-    case instance;
+    /**
+     * @param ?Sequence<Period> $retries
+     */
+    private function __construct(
+        private ?Sequence $retries,
+    ) {
+    }
 
     #[\NoDiscard]
     public function __invoke(Config $config): Config
@@ -31,6 +39,15 @@ enum Resilient
     #[\NoDiscard]
     public static function new(): self
     {
-        return self::instance;
+        return new self(null);
+    }
+
+    /**
+     * @param Sequence<Period> $retries
+     */
+    #[\NoDiscard]
+    public function withHttpRetries(Sequence $retries): self
+    {
+        return new self($retries);
     }
 }